refactor: Session 승인/취소 로직 조건 추가
